implemented untested opcodes

+++ cpu/cpu.h
@@ -14,7 +14,7 @@ typedef struct stack{
 typedef struct chip_8_machine{
   uint8_t *chip_8_memory;
 
-  int pc_counter;
+  uint16_t pc_counter;
   uint16_t pc_counter_end;
   uint8_t* game_start_address;
